Improve del.py

+++ ecmascript/transforms/scripts/del.py
@@ -1,13 +1,14 @@
 #!/usr/bin/python3
 
 
-from os import listdir, remove
-from os.path import isfile, join
 import shutil
+from os import listdir
+from os.path import isfile, join
+
 
 def search(dir):
     print('Searching ', dir)
-    if len(listdir(dir)) == 1:
+    if len(listdir(dir)) == 1 and 'fixtures' in dir:
         shutil.rmtree(dir, ignore_errors=True)
         return True
 
@@ -20,9 +21,11 @@ def search(dir):
         else:
             if not search(p):
                 shouldDelete = False
-    if shouldDelete:
+            if isfile(join(dir, 'options.json')) and not isfile(join(p, 'options.json')):
+                shutil.copyfile(join(dir, 'options.json'), join(p, 'options.json'))
+    if shouldDelete and 'fixtures' in dir:
         shutil.rmtree(dir, ignore_errors=True)
     return True
 
 
-search('./fixtures')
+search('./')
